requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
bio==1.5.9
biopython==1.81
biothings-client==0.3.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
django-mysql==4.9.0
djangorestframework==3.14.0
dynaconf==3.1.12
filelock==3.9.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gprofiler-official==1.0.0
grpcio==1.56.0
h5py==3.9.0
hvac==1.1.1
idna==3.4
jax==0.4.13
keras==2.12.0
knox==0.1.14
libclang==16.0.0
loguru==0.7.0
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mygene==3.2.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
path==16.6.0
Pillow==9.4.0
platformdirs==3.0.0
pooch==1.7.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyhcl==0.4.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2022.7
urllib3==2.0.2
validators==0.20.0
virtualenv==20.19.0
Werkzeug==2.3.6
win32-setctime==1.1.0
wrapt==1.14.1

